+++ packages/flutter_riverpod/test/providers/change_notifier/change_notifier_provider_test.dart
@@ -5,7 +5,6 @@ import 'dart:async';
 import 'package:flutter/widgets.dart' hide Listener;
 import 'package:flutter_riverpod/src/internals.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
 
 import '../../utils.dart';
 
@@ -75,21 +74,6 @@ void main() {
     expect(container.read(provider.notifier), null);
 
     container.dispose();
-  });
-
-  test('can read and set current ChangeNotifier', () async {
-    final container = ProviderContainer.test();
-    final listener = Listener<ValueNotifier<int>>();
-    late Ref ref;
-    final provider = ChangeNotifierProvider<ValueNotifier<int>>((r) {
-      ref = r;
-      return ValueNotifier(0);
-    });
-
-    container.listen(provider, listener.call);
-
-    verifyZeroInteractions(listener);
-    expect(ref.state.value, 0);
   });
 
   test('can refresh .notifier', () async {