+++ packages/riverpod/test/old/legacy_providers/deprecated/state_notifier_provider/state_notifier_provider_test.dart
@@ -90,21 +90,6 @@ void main() {
     verifyNoMoreInteractions(listener);
   });
 
-  test('can read and set current StateNotifier', () async {
-    final container = ProviderContainer.test();
-    final listener = Listener<int>();
-    late Ref ref;
-    final provider = StateNotifierProvider<Counter, int>((r) {
-      ref = r;
-      return Counter();
-    });
-
-    container.listen(provider, listener.call);
-
-    verifyZeroInteractions(listener);
-    expect(ref.state, 0);
-  });
-
   test('can be auto-scoped', () async {
     final dep = Provider(
       (ref) => 0,