+++ packages/riverpod/test/old/legacy_providers/future_provider/future_provider_test.dart
@@ -39,44 +39,6 @@ void main() {
     expect(container.read(autoDispose).value, 84);
   });
 
-  test('Does not skip return value if ref.state was set', () async {
-    final provider = FutureProvider<int>((ref) async {
-      await Future<void>.value();
-      ref.state = const AsyncData(1);
-      await Future<void>.value();
-      ref.state = const AsyncData(2);
-      await Future<void>.value();
-      return 3;
-    });
-    final container = ProviderContainer.test();
-    final listener = Listener<AsyncValue<int>>();
-    // Completer used for the sole purpose of being able to await `provider.future`
-    // Since `provider` emits `AsyncData` before the future completes, then
-    // `provider.future` completes early.
-    // As such, awaiting `provider.future` isn't enough to fully await the FutureProvider
-    final completer = Completer<void>();
-
-    container.listen<AsyncValue<int>>(
-      provider,
-      (prev, next) {
-        if (next.value == 3) {
-          completer.complete();
-        }
-        listener(prev, next);
-      },
-      fireImmediately: true,
-    );
-
-    await completer.future;
-
-    verifyInOrder([
-      listener(null, const AsyncLoading<int>()),
-      listener(const AsyncLoading<int>(), const AsyncData(1)),
-      listener(const AsyncData(1), const AsyncData(2)),
-      listener(const AsyncData(2), const AsyncData(3)),
-    ]);
-  });
-
   test('supports family overrideWith', () {
     final family = FutureProvider.family<String, int>((ref, arg) => '0 $arg');
     final autoDisposeFamily = FutureProvider.autoDispose.family<String, int>(
@@ -209,38 +171,6 @@ void main() {
       await expectLater(container.read(provider.future), completion(1));
       expect(container.read(provider), const AsyncData(1));
     });
-  });
-
-  test('can read and set current AsyncValue', () {
-    final container = ProviderContainer.test();
-    final listener = Listener<AsyncValue<int>>();
-    late Ref ref;
-    final provider = FutureProvider<int>((r) {
-      ref = r;
-      return 0;
-    });
-
-    container.listen(provider, listener.call);
-
-    expect(ref.state, const AsyncData(0));
-    verifyZeroInteractions(listener);
-
-    ref.state = const AsyncLoading<int>();
-
-    expect(
-      ref.state,
-      const AsyncLoading<int>()
-          .copyWithPrevious(const AsyncData(0), isRefresh: false),
-    );
-
-    verifyOnly(
-      listener,
-      listener(
-        const AsyncData(0),
-        const AsyncLoading<int>()
-            .copyWithPrevious(const AsyncData(0), isRefresh: false),
-      ),
-    );
   });
 
   test('can be auto-scoped', () async {